Reorganized source directory into top-level unit, ui, and util packages; related setup and test updates.
