+++ test/unit/test_generate_execute.py
@@ -23,10 +23,10 @@ import unittest
 import copy
 
 sys.path.insert(0, os.path.abspath(os.path.dirname(__file__))+os.sep+'..')
-from tkltest.generate.unit import generate
-from tkltest.execute.unit import execute
+from tkltest.unit.generate import generate
+from tkltest.unit.execute import execute
 from tkltest.util import config_util, constants
-from tkltest.util.unit import dir_util
+from tkltest.unit.util import dir_util
 
 from pathlib import Path
 import xml.etree.ElementTree as ElementTree
