+++ tkltest/ui/generate/ranked_attributes.json
