+++ tkltest/ui/generate/ranked_attributes_form_fields.json
