+++ tkltest/ui/util/browser_util.py
