+++ tkltest/unit/execute/execute.py
@@ -22,11 +22,11 @@ import shutil
 
 import toml
 
-from tkltest.execute.unit import compare_coverage
+from tkltest.unit.execute import compare_coverage
 from tkltest.util import constants, command_util, config_util
-from tkltest.util.unit import build_util, dir_util, coverage_util
+from tkltest.unit.util import build_util, dir_util, coverage_util
 from tkltest.util.logging_util import tkltest_status
-from tkltest.execute.unit.coverage_html_writer import CoverageStatisticsHtmlWriter
+from tkltest.unit.execute.coverage_html_writer import CoverageStatisticsHtmlWriter
 
 
 def process_execute_command(args, config):
