+++ tkltest/unit/util/code_util.py
@@ -15,7 +15,7 @@
 # ***************************************************************************
 
 
-from tkltest.util.unit import java_class_parser
+from tkltest.unit.util import java_class_parser
 
 
 def get_methods_lines(class_file_name):
